feat: add vica integration
The vica chatbot integration requires not just the inclusion of
chatbot scripts (which we can already insert through an existing
include) but also the insertion of a stylesheet in the head of the
website.

Since we don't currently have a suitable way of adding to the
head of the website, this commit adds a new `site.vica` option which
if true, includes the vica stylesheet as part of the website head.